Adicionado novo endpoint para consulta de contato por id e criado nova entidade email

+++ src/main/kotlin/com/adriane/start/kotlin/repositories/ContactRepository.kt
@@ -3,7 +3,7 @@ package com.adriane.start.kotlin.repositories
 import com.adriane.start.kotlin.entities.Contact
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ContactRepository : JpaRepository<Contact, Long> {
+interface ContactRepository : JpaRepository<Contact, Int> {
 
     fun findByActive(active : Boolean) : List<Contact>
 }
